+++ noodles-vcf/CHANGELOG.md
@@ -6,6 +6,8 @@
 
   * vcf: Implement `std::error::Error::source` for errors.
 
+  * vcf/record/info: Implement `Extend<(Key, Option<Value>)>`.
+
 ### Changed
 
   * vcf/header/builder: Change inserting ALT, contig, FILTER, FORMAT, INFO, and